Update id of object with id from database
The Simulation objects in Python have a sim_id,
which is used to identify the simulation when
printing various status updates in the server.

The Simulation rows in the database also have an id,
which is automatically assigned to them on insertion.

This commit uses the database ids as the object ids.